backend/requirements.txt
fastapi>=0.116,<1.0
uvicorn[standard]>=0.35,<1.0
httpx>=0.28,<1.0
google-genai>=1.46,<2.0
